backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.0
psycopg2-binary>=2.9
anthropic>=0.25
python-dotenv>=1.0
gunicorn>=21.0
